fix(B-464): move paginated session schema to schemas folder
